Updated tilemap collision and character movement logic
By adding a composite collider to a tilemap, it can create a collider that spans tiles and gets rid of the clipping bug that appears with a box collider between two tiles.

Added some logic for crouching where the player cannot uncrouch in a tight space where the full size model would not fit. It is dependent on the animation length for roll and slide. Was using magic numbers for debugging purposes but these will eventually be variables.