Change ground/uncrouch checking
Using trigger colliders to do the checking now. This way is more optimized since it no longer does a check every frame. Also, the player's collision detection for its rigidbody2d component was changed to continuous from discrete. This fixes a lot of weird issues with Unity predicting the player being somewhere it cannot be.

Still have to optimize the CharacterMovement script for ground/uncrouch checking and add logic for attacking.