Added more trigger boxes on the character
Added some trigger boxes for the player: one for interacting and two for attacking. The implementation in scripts has not been added yet.

Still need to optimize CharacterMovement script so that uncrouch/grounded checks are only done when the collider does not overlap with anything.

Changed the FaceLeft and FaceRight animations to change the character's scale instead of sprite. This will also change the direction of all of the trigger colliders where it would not before.