Added Anima2D and basis for interaction.
Anima2D is a Unity plugin that makes animation a lot easier. I've added it to our project for future purposes.

I've created a basic interaction system where items have an interaction type. If one item is in the interaction range, then it is interacted with. If more than one item is in range, however, a menu pops up and the player can choose the item that they want to interact with. This menu system can be adapted to virtually every other menu.