Small code change and ui adjustments
Fixed a comment in the InteractionManager script and adjusted the user interface for the health bar and interaction list. The health bar is now red and the interaction list is anchored to the left side of the screen for now.

The end goal is to anchor the interaction list to the side that the player is not facing (since the stuff they are interacting with is always in front of them).