Removed Anima2D and added level streaming
Since we are now going with spritesheets for animation, Anima2D is no longer needed in the project.

I've added a level streaming system which will unload/load two adjacent levels seamlessly. All of game objects not related to an individual level (i.e. the player, canvas, and event system) have been moved to a new, persistent level. This level contains a bunch of hallways which will connect each of the individual levels in the game.

Since these hallways are pretty small, the performance impact should be minimal. These hallways also contain a loading/unloading mechanism consisting of 6 trigger colliders. The outer two trigger colliders unload the level on the opposite side if it is not already unloaded. The next two interior colliders load the level on the opposite side if it is not already loaded. The middle two colliders ensure that both levels are loaded before the player continues. If not, then the player will have to wait until they are loaded. This part has not been implemented yet.

The next update from me will move all of the input from the player to one function in order to generalize the movement/attack scripts so that they can be used for AI as well.