Added knight animations, camera fix, and movement fix
Added the spritesheets that Dom made for idle and walk movement of the knight.

Fixed the camera to where it is no longer a child of the knight object. Now, it tracks the knight with some smoothing.

Fixed the movement inconsistency at different framerates. The movement multipliers were being set multiple times per frame at low framerates. Input is gathered in Update and movement is acted upon in FixedUpdate. At lower framerates, FixedUpdate will happen multiple times per update, so the multiplier effects were two/three/four fold in these scenarios. I changed the multiplier to not affect the movement directly so that it can't stack like this.

Besides that, there was some refactoring with the player input in PlayerInput.