Refactoring, Layers, and Optimizations
Refactored pretty much all of the existing scripts and moved them into a folder that makes more sense.

Added layers and adjusted collision between those layers. The player, enemies, and interactables no longer collide with each other. All components off of a character or enemy have been given the layer ChildComponent which allows them to not collide with each other.

Optimized the CharacterMovement script a bit so that it does less work per tick. Now, grounded checks no longer happen in update.

If a character is walking/running into a wall, they will still be in the idle state. This is accomplished via a position check every tick. If the position has not changed, then the player is not moving.