Trigger Abstract Class, Refactoring, and Attacking
Made an abstract class Trigger. This class is inherited by all scripts that were keeping a list of objects in a trigger collider. This way, we are duplicating less code. There is an abstract function in the class called TriggerAction which needs to be implemented by each derived class. This is where you put the implementation for each individual trigger.

Refactored all of the trigger classes to derive from the Trigger class. Renamed all of them from <name>Manager to <name>Trigger to represent what they do better. Removed the InteractionManager and put all of that code in CharacterInteraction. Renamed InteractTrigger to InteractionFire since it fires interaction events (and because we now have an InteractionTrigger class for the interaction trigger box).

Added a CharacterAttack class which will allow for a character to attack if they have input to attack and there are characters in the trigger box of the attack.

We need to figure out how to set up the attacks. Do we want 3 different attack buttons or to have attacks based on movements? On one side, it gives the player more control over which attack they want to use with separate buttons. On the other side, it makes the player input more complex which may degrade the user experience.