Added data elements and player animations
Started working on the data elements of the project.

GameData will hold all of the data objects in one place. This includes data for dialogue, characters, items, and the player. The player data object will hold all necessary information to load the game from memory.

Character database is an XML file that will be accessed by the script CharacterData. The same will be true for ItemData.

DialogueData will exist as a set of XML files: one key file which will point to the specific XML file of a conversation. An interactable character may have multiple dialogues (read: XML files) and there will be many characters in the game.

PlayerState will be written to any time a significant event happens in the game and read from when the game is loaded. There may be several files to load from if the need arises.

From the main menu, we will also need a data file that accesses each of the game state files (PlayerData) and the game configuration from the options menu.

Brought in the spritesheets for jump, fall, and slide. In terms of spritesheets for the player, we still need run, crouch, crawl, roll, wind_up_attack, swing_attack, stab_attack, and power_attack.