Character and item data objects
Turned each data object into a ScriptableObject since it did not need to exist as a MonoBehaviour.

Created sample item and character databases which are loaded into memory at runtime. These can be accessed via the GameData object within the game and contain all of the necessary information about items and characters respectively.