Started Beacon System and Boxcasting
Beacon System info:
Have the setup for the beacon seystem but implementing it is not
working. May have to scrap and start from the beginning.

BoxCasting:
All that is need for it is for the world to be casted and the enemy
itself to be ignored in the cast. Layermask was looking like a good
solution but it either removes and doesnt track world or tracks world
and enemy but not player.