Working Inventory UI
The player can pick up items that exist in the world and they will be added to their inventory. Right now, the inventory is accessible via the "R" key. Items can be selected to view more information about them but they cannot yet be dropped/used/destroyed.

I also created the skeleton for the backpack menu (inventory, journal, and map).