Added animations, fixed player colliders, and fixed backpack UI issue
Added the crawl and crouch animations to the player. The stab animation will be added once power, swing and wind-up are added (since they need to be added to the state machine). Besides that, the only other movement animation that we need is roll.

I also adjusted all of the player colliders so that they make sense with the knight's sprite (they used to be based off of the placeholder squares used previously). With how it is right now, the player's arms and part of the head go in front of walls. In my opinion, that is preferable to having an oversized collider for the player.

We may at some point want to instead derive the location and scale of each collider mathematically with respect to the main collider so that we do not have to adjust a bunch of colliders every time we make a change to a character.

Fixed a UI bug where the scrollbars would be inaccurate if the resolution or aspect ratio was changed. Instead of setting the highest point of the ScrollRect container at instantiation, I just set the ScrollRect back to the top using a member variable.