Added attacks, fixed player sprite, adjusted player colliders
Added a coroutine which performs the stab, swing, and power attack. This coroutine runs for the duration of the attack (or it should at least-- some tuning may be required).

Modified the animation controller for the character to include attacking and winding up.

Fixed the player sprite to be 128x128 in all cases with the player centered.

The only animations we need for the player at this point are the power attack, winding up while running, winding up while walking, winding up while idle, and a death animation.

Other possible animations include a greeting animation (a wave or something when entering a dialogue with another character) and an interaction animation (picking up/using something).

I still need to prevent attacking while not in idle, walk, or run and vice versa.