Big UI update with functional inventory
Recreated all of the UI elements to anchor to a proportion of the screen so that the UI now scales to any resolution. For example, the backpack menu is always three quarters of the width of the screen regardless of the resolution. All of the elements inside the backpack (tabs, items, description information, etc.) are set up the same way.

In the inventory in the backpack menu, I left some room at the top-right for stats such as player health, attack, defense, speed, gold, and the currently-equipped weapon/armor.

The interact menu now has a "collect all items button" which is located right next to the exit button in the top right-hand corner. It will pick up all items in range of the player (i.e. all items in the menu).

Items can now be used, dropped, and destroyed.

Using an item will send output to the debug screen. Once we start wiring up weapons, armor, and health I will make each item do what it is supposed to do.

Choosing to drop an item will bring up a dialogue box which asks how many items the player wishes to drop (one, all, or cancel). Dropping an item will instantiate it in the world at the player's location. The instantiation occurs in the persistent level. If the player leaves the area and despawns the level, the item will fall to a collider which destroys it. If the item is dropped in one of the hallways that connects levels, it will exist there until the game is quit.

Destroying an item is the same as dropping it without instantiating it. However, an extra dialogue box pops up to confirm the action since these items are unretrievable after destruction.

I moved the interact bar ("Press <key> to <action> <item>") onto another canvas which scales to camera space. I plan to create a world space canvas for each enemy in the game so that their health is shown above their head.

Refactored the "Fire" scripts a bit which call a function on a UI event.