Added door interaction, enums for interactable types, and moved interact bar
Added interaction for door interactables. There are two types of doors: enter doors and exit doors. The only way they differ is by the text "enter" or "exit".

The InteractableObject class now has enums for the types of interactable items: EnterDoor, ExitDoor, Character, Chest, and Item are the current types right now.

The Door class inherits from the InteractableObject class. It includes the paired door that it teleports the player to upon interaction and a coroutine to carry out the animation and delay before entering.

Moved the interact bar down a bit given the new camera location. I think the camera may be a bit too high which makes it hard to see things below you.