Added some interactable scripts and side doors
There are now two types of doors: back doors and side doors.

Back doors are the type of doors that existed previously. They will teleport the character to a different place.

Side doors are doors that simply open and close. When closed, they block the player's path.

Renamed the InteractableObject class to Interactable for simplicity's sake.

Added animation controllers and prefabs for each type of door.