Moved Cinemachine, added group item collection, and added particle effects
Moved Cinemachine into the plugins folder with the other plugins.

When the player hits the collect all items button in the interact list, now every item lerps towards the player (thanks Aris).

Added a particle effect gameobject as a child of the player. Currently, the only particle effect is for the potions (a little green heal effect that triggers whenever the player uses a potion).