Started refactoring character classes
I didn't like the organization of the Character scripts, so I am refactoring them a bit to give them an inheritance structure that makes more sense and is much more clean. Another reason to do this is so that each character type only has the necessary behaviors and fields (e.g. NPCs do not need to have any information about attacking or crouching).

Basically, Player and Enemy derive from ActiveCharacter (characters that attack). PassiveCharacter includes characters that do not attack and only exist for dialogue purposes. Both ActiveCharacter and PassiveCharacter derive from Character.

This above setup will apply for attacking, interacting, and moving. Only the player can interact and crouch (and climb possibly). Only active characters can attack (but the way that an enemy attacks is much different from the way the player attacks (enemies rely on AI, player relies on user input).

I'll give more information on the way the classes work together once more progress has been made. The player works properly with the changes, but the enemies will be broken for a little while.

Besides that, I reorganized the animation and scripts folder a bit.