Climbing, Attacking, Refactoring, LevelActivating, and more
Added climbing to the player movement. When the player is overlapped with a ladder, they will be allowed to climb it. While on the ladder, the player can strafe left or right, hang idle, and climb up or down. Ladders pair particularly well with platform effector colliders with one-way collision. With these, you can climb up through a collider and be standing on a platform.

Added knockback and red tint to a target upon a successful hit of an attack.

Changed the attack logic so that it checks for a new enemy every tick while attacking. Previously, it would inflict damage in only one frame of the attack. This is bad because if an enemy enters the attack trigger after the frame has passed, the hit will not be counted. Now, I am keeping a list of all of the gameobjects that have been hit for a given attack.

Made much more progress into refactoring the character classes. The player-related stuff has been completely refactored. Pretty much all that is left is refactoring the enemy attacks.

Added more logic to prevent the player (and characters in general) from performing actions from certain states. For example, while a character is attacking (and while the player is interacting), their movement updates are disabled.

Added the ability to disable player input. The boolean acceptInput in PlayerInput disables all movement-related input during certain events like when the interaction or backpack menus are active.

Changed the level streaming logic a bit. The outermost triggers now activate the level closest to the trigger. The next two outermost triggers activate the persistent level. The center two level manager triggers now disable the player input if the respective level has not loaded fully. It reenables player input once loaded.

Updated the Animator (state machine) to include two interaction states (one while crouched and one while standing) and three climbing states (climb, climb-strafe, climb-idle). Removed a lot of the transitions from the Animator to simplify it and fixed a couple of transition issues (one of which would not allow transitioning from one wind-up state to another).

Added a climbing trigger which keeps track of ladders for the purpose of climbing. It behaves similar to the other triggers (grounded, stabbing, etc.).

Player animations needed: idle-wind-up, walk-wind-up, run-wind-up, interact, interact-crouch, climb, climb-strafe, climb-idle, and death.