Fixed Camera Shifting and Background Scrolling
Made it so that the camera shift works in pretty much every scenario. Added a zooming feature to it so that the camera gets up close and personal with the player character.

If the player activates the backpack when they are located on the left side of the screen, it will snap the backpack to the right side of the screen and zoom in on the player to the left. Vice versa for the right side.

The last thing that I might work on with the zooming is determining the optimal zoom based on the player's size relative to the screen.

Changed the rest of the animations to prevent interacting in states where the interact animation would not play correctly.

Changed the background scrolling so that it is a function of the camera's velocity instead of the player's. This prevents background scrolling when the camera is not moving so that the background only moves when the foreground does.

Added some movement smoothing to the CharacterMovement class so that they cannot exit a motionless state at full speed. It is still a work in progress.