DialogueData, dialogue container, and dialogue classes
DialogueData now loads all game dialogue into memory when the game starts. Dialogue is located in sandbox\Dialogue. Each directory inside is a character/object which the player can enter dialogue with. The name of the directory is the name of the character. Inside a character's folder are all of the conversations the character can have. These conversations are stored as XML files as a set of conversation segments.

The dialogue is stored as nested dictionaries.

In the outer level, each dictionary entry is a character where the key value is the unique name of the character/object.

In the next inner level, each dictionary entry is a conversation where the key value is the id of the conversation. The conversation ids are kept in the filename of the XML conversations. These filenames can be appended with a string to identify the file as long as the integer id and string are separated by an underscore (ex: 1_Greeting.xml).

In the next inner level, each dictionary entry is a dialogue segment where the key value is the unique id of the segment. From this point in, it is the same format as it was previously with the exception of the actions.

Dialogue actions are events that occur as a result of a certain path taken in a dialogue. It is stored at the same level as a dialogue choice. One segment may have multiple actions which are kept as strings and turned to their enum equivalent at the start of the game. Some examples of actions include Disappear, BecomeHostile, RequestGold, GiveGold, RequestItem, etc.

Added a new pixel font which has the Desktop license. I like this one more than the current font because it has lowercase and each letter is one continuous piece (not separated as with the highway sign font).

Adjusted the player attack to not occur when the player is falling. If the player starts to fall, the attack will be cancelled.

Brought in the dialogue to Persistent-SC on the character standing on the platform. This character is named villager and its dialogue state is 1 (i.e. it will carry out its conversation with id 1).

Created a new container for the dialogue which is pretty similar to the interact list. It takes up half of the screen.

Disabling player input now disables the interaction trigger on the player in all cases except when the interaction list is up (since it depends on the items located in the interact list).

Added an XElementExtensionMethod for the string actions which are parsed to enums.

Changed the dialogue segments to not have a type. Instead, the amount of choices is dynamic. If there are 0 choices, it will show a continue button. Otherwise, it will show all of the choices and allow the player to choose one.

Storing the dialogue data in dictionaries is much more efficient on time. Instead of having to loop through the entire list looking for a value, you can enter the key and that will just outright give the value.

Simplified the InteractionPrefabReference class to just store the index of the interact button in the list and return that index if selected instead of looping through all of the interactables and finding the one with the selected text.