Player climb/strafe animations, goblin animations, HSV shaders
Added the player climb and strafe animations. Changed the animation controller for the player to transition to climb-idle from walk/run/jump/fall/idle instead of from any state because it was transitioning to climb-idle from climb-idle which prevented the animation from being carried out.

Added the goblin animations and simplified its animation controller to not include crouching, climbing, or interacting states. Also, it only has one wind-up state because enemies are only going to be able to wind up while they are not moving.

Added a cool HSV shader which allows for us to change the color of the sprites through a material and choose which colors of the sprite to apply the shader to. The end result is that we get vibrant possibilities of colors for each character without having to deal with discoloration side effects we would have had with Unity's RGB hue system.