Added dialogue actions and adjusted interact trigger
Added a couple more dialogue actions and put some examples in Villager's 1st conversation. When a segment which has actions is encountered, it will perform the actions at the end of the segment processing (i.e. right after you choose a choice or hit continue). It will perform all actions in the list.

The actions themselves still need to be added, but that will be done after more work has been done with the saving/config.

Adjusted the interact trigger to be in control of enabling/disabling its box collider. On enable, it shifts/wobbles its collider's size a bit so that it will notice any items which were inside on disable. The only case which it does not pick up now is after a power attack.