Fixes for climbing, attack, and dialogue
Fixed climbing so that the player no longer can fall off the ladder when they get close to the top. I added another climbing trigger a little higher on the character. When the player reaches the top, I do a check to see if it that trigger is empty. If it is, don't exit the ladder until the bottom one is empty. These ladders cannot be climbed back down after the top is reached.

Added two functions to active character attack: one that's carried out before an attack and one after. These are overridden in the player attack script to disable/enable player input. With this, I fixed the issue where you could not interact with anything after doing an attack (unless you moved a bit to let the interaction trigger know that there is an interactable inside).

Made some small changes to the way that dialogue is processed. After a dialogue is completed, the choices are flushed from the container. They were kept there previously which caused a couple of issues (actions in the last segment would be repeated each time I hit spacebar). Also, I moved the choice instantiation to a function which is called after the dialogue text is finished animating out letter by letter.