Autoclimb
Tightened the trigger boxes for climbing on the player. Now, they can only climb when about half of their body is on the ladder.

If the player nears the top of the ladder (and their feet are touching the platform at the top of the ladder), the rest of the ladder will be automatically climbed up. Since the player cannot climb down in this scenario anyways (cannot climb down when grounded), it is best to send them up the rest of the way to avoid confusion. The thinner the platform is at the top of the ladder, the less dramatic the autoclimb is.