Roll force, crouch fix, interact trigger fix
Added a force upon rolling so that the roll may be used as a dodge attack or a quicker way to get a long distance into an area that can only be crawled through.

Fixed crouching so that if the player interacts while crouching they do not exit the crouch state. If the player interacts while crouching, it resets the timer in which they can uncrouch. Since this timer may be reset, I added a boolean value to only apply the bonus speed while sliding as the state is entered (instead of after an interaction while crouched).

Fixed the interact trigger. Its size would change more than it was supposed to since the change was being carried out in a coroutine. If the coroutine did not finish, the trigger would never be reset to its default value. Now, it sets the trigger back to the default value before and after changing its size.