Castle Sprite is in
Created the castle sprite, I am currently working on a door for the
castle, it has to be Majestic so I am taking time.

I also changed the background for the Level and the way it is now will
probably be the Alpha version.

Created and Added the castle wall background.

Changed some Camera things again (mostly the collider)

Need to find a way to make the platform tile only take up half the cube.