Snap fix when entering doors
When entering doors, I disable the vcam and enable it once the position is updated. This fixes the issue with the camera spazzing out.

I was messing around with Cinemachine so I do not want to push the updated level. You will have to replace the Character script on the player with a Player script and set the references for it to run.