Reorganized scripts, dialogue zoom, fixed camera shifting, interact+windUp animations
Reorganized the scripts into folders. The main menu scripts have been moved to the main menu folder, and the camera scripts have been moved to the camera folder.

Added zoom for dialogue where it zooms to the left or right half of the screen depending on the screen space of the character. For the zoom, instead of just lerping the orthographic size, I am now smooth stepping the orthographic size and x offset of the camera. When I zoom in, I set the x damping to 0 and when I zoom out I set it back to its default value. The result is that the camera shifts directly where I need it to without overshooting.

I added a virtual camera to my scene without the polygon collider and set the dead zone to a pretty low value (0.1).

Added the interact animations to the player for idle and crouch. Added the climb idle animation as well. For the regular interact state, I made transitions to it from run/walk/idle instead of any state since it was not letting it play through (transitioning to itself once again).

Added the wind-up animations. We may need to add a frame in front of the stab animation as it looks a bit awkward with the transition now.