Better camera shifting and interactable fix
Added some more variables to the coroutine so that the shift is even smoother than it was before. Now, camera offset (x and y), dead zone (width and height), and orthographic size are all smoothed.

Also, I fixed the health potion in level 2 to be a pickup.

To prevent future issues with items in the future, I made it so that Interactable is an abstract class and the type is not serialized for the editor (to allow the type to be edited). Each type is set in the Start function of its derived class.