Implemented chests, Game data singleton, updated sorting layers
Brought in the chest, created an animation for it, and added the event that takes place upon interacting with it (spawns a specified item).

Updated the GameData class to be a singleton. This ensures that only one instance of all of the game data will be active at any time. Instead of referencing the data via game object passing, I made a static instance of the class in itself that can be accessed anywhere else in the project.

Changed the way sorting happens for interactables. Now, there are BackgroundInteractables (doors, chests, NPCs) and ForegroundInteractables(items) so that items always appear in front of other interactables.

Started working on a unique ID for each interactable in the scene. Since the state of interactables persists between scenes, it is necessary to know things like whether a chest or door is open/unlocked or not. The unique ID is less important for items.