Controller support and (more) controller-friendly UI
Added controller support for Dualshock 4 controllers. Xbox (1 and 360) controllers should work as well, but the button mapping will be off. For the full game, we can specify button profiles for certain controllers and the player can just select the type of controller that they are using.

To make the UI work with controllers, I added a class which sets the focus to a specified UI element (i.e. the first button in the interact menu/dialogue menu/inventory). In order to get the ScrollRect to scroll with the currently-selected item, I had to get its index under the ScrollRect and set the ScrollRect's position off of that.

Each class which brings up a UI passes its ScrollRect, list under the scrollRect, and first item in the list under the ScrollRect to the ElementFocus class which takes care of focusing on the current element and keeping it in view.

Added the updated icons to the inventory menu.