Wraith walking sheet and other stuff
Reorganized the tile sheets because I made two of the Castle ground
tiles.

I also created the Castle scene to help with testing right now it is
connected to Area 1 (need to open scene additive) but I plan to separate
them so we dont have camera issues.

The Wraith is now animated for walking. Currently trying to think of
three "attacks" and the death. Running will just be a faster version of
the walk.

Added a Castledoor animator but cannot figure out how to add the castle
door sprites. Should also mention that I split the castle door sprites
into 2 images as well  (I still need to fix their settings).