AI is now walking and jumping over objects
AI in my testing level is now walking and jumping over obstacles, as of
pushing this there is a possibility that the AI will jump over the
player but I will double check this later.

The AI walks in its set path using a Beacon system, to use it just place
to beacons somewhere(Prefab) and then assign the enemy to said beacons.
I will probably change this to an array so multiple enemies can use the
same Beacon.