Added more ground tiles, and updated background
Added more to the castle level, the kings throne room is in, just need
to spruce it up a little to make it look pretty.

THERE ARE TEMP BACKGROUNDS IN RIGHT NOW
The temp backgrounds will be removed once the parrallax background
switching works.

I also need to remove the ground from castle to persistent.