BetterBoxes™
The boxes now behave much better. Composite colliders and OnTriggerExit2D do not behave well together so that led to a couple of issues that needed to be fixed with both the player and boxes.

The boxes have several trigger components on them that determine whether they can be moved or not depending on whether or not they are touching another box, the player is touching them, or they are falling.