Project Reorganization
Reorganized the entire project so that it isn't as much of a mess. Let me know if you can't find something that you are looking for.

Resized the sprites so that they are only as large as they need to be (most were reduced from max2048 to max256 or lower).

Got rid of the starter tiles that I brought in early on.

Refactored the CharacterMovement and ActiveCharacterMovement into one class called CharacterMovement. Since the NPCs won't be moving, there is no need for a movement script for them. However, enemies may have an NPC script on them in which case they can become hostile.

Added in the lock state for all of the doors and chests and repositioned the center of the regular doors (player was teleporting into the ground).

Renamed the animation states to follow a naming convention.

Added a class Lockable which will be used to lock the chests and doors until some action is performed. The action could be destroying a set of gameObjects, having a key, pressing a button, or a set of actions for the tutorial (stab/swing/power hit the dummy three times).