Revamped attacking
Attacking now consists of one trigger: AttackTrigger. Each enemy can register up to three unique attack types which correspond with an animation. That animation sets the bounds of the attack trigger and when exactly it is enabled and not enabled.

The player has the attack trigger on the weapon itself so that the animations can be changed depending on the weapon type.

Enemies will simply wind up for a period of time (where they display a particle effect that corresponds with the attack that they have randomly chosen) and then pick the attack.