Started implementing saving and added sqlite
Saving consists of a bunch of events which will write to a sqlite database. Whenever something of importance happens, an event will be raised which fill fire a function in the SaveData class with all of the information it needs to update the database to reflect the current game state.

Added the necessary sqlite dlls so that we can build for windows, mac, and linux.

Adjusted the hitbox on the boxes slightly so that they do not get stuck on their way down into a hole.

Fixed the castle so that it is no longer using the spooky material.