More tiles, and orthographic zoom
Added some more slanted tiles nothing major.

Added a orthographic camera shift using a trigger. The triggers have to
be stored in persistent. Also the shift time and shift amount are values
on the trigger themselves so we can shift in and out differently in
different areas, the triggers themselves need to cover the whole area
you need to shift on.