Added buttons, UnlockActions, SQLite saving, and openables
Buttons can be pressed to open doors/chests or reset a set of crates. Buttons can be toggled to need to be held down or stay pressed once initially pressed.

UnlockActions are are things which unlock doors/chests upon a set of criteria being met. The four actions currently are:

-DestroyGameObjects, where the player must destroy a set of GameObjects.

-ButtonsDown, where a set of buttons must be pressed down.

-CheckPoint, where a door must be entered.

-HaveKey, where a key must be in the player's inventory.

SQLite saving has been added for doors and chest states.

Openables is a class which SideDoor, BackDoor, and Chest derive from. This makes it so only one event has to be triggered from the openables class instead of three (one from each derived class). This is also the place where the checking for locking/unlocking happens for the door/chest based on the UnlockAction that is on the door/chest.