Exploding tiles, attacking bear, respawning, spawn manager
Added the exploding tiles in the BearChase scene when the bear destroys the tiles. They spawn, play their animation, and then get destroyed.

The bear now attacks whenever a block is destroyed if it is not already attacking. It is just set to an attacking state so that it does not have to stop in place.

Made some slight optimizations to player respawn where they respawn at a persistent level location and have their velocity set to zero.

Created a spawn manager at each level loader that sets the player's spawn upon entering it. If the player dies, they are respawned at the persistent location and the levels on either side are refreshed (unloaded if they are loaded and then reloaded). I added a loading prompt to display when this happens.

Adjusted the KillY so that it does not destroy the player or any components on the player. Instead, it just does enough damage to kill them. The bear does the same.