Optimizations and collision fix
Optimized the animations to not loop when not needed to as well as not play unless the camera is on them.

Brought in an updated version of TextMesh Pro.

Updated the tiledestroyer to ignore the platform layer as the bear was destroying the platform tilemap. Also, on the bear, I set its platform effector to ignore the platform layer so that it would not collide with it.

Made the buttons which respawn crates blue to differentiate them from the red crates.

Added all of the assets to an assetbundle which is loaded at the start. I'm not sure this is affecting the performance either way, so I may need to configure it more. After profiling the lag spikes in the level loading I found that it was almost entirely the BuildPipeline causing the spike. Even more interesting is that the spike does not happen every time. Sometimes the level loading will work properly and load over time instead of right upon entering the load trigger.

Got rid of threadninja as it could not be used with level loading. There was not really a point for it besides that.

Created a prefab for the level loader and consolidated the data entry to the child objects _LeftLevel and _RightLevel.

Created a singleton in the Player, PlayerInput, and BackgroundSwitch classes so that a prefab could be created for the level loaders. These classes will only and always have one instance anyways.

I'm taking a look into Zenject to see if it is feasible for the project.

Made some small changes to the level layout of BearChase-AP. Added ladders and doors.