Animator Override Component
Instead of using a new animator controller for each and every character in the game, we can specify an animator controller to override with a different set of animations using the animator override component.

Here, we just specify the set of animations we are using for this specific character and those will take the place of the defaults. For example, the Character_AnimationController is the animator controller which will be overridden by each enemy in the game. Whenever we bring in a new character, we will set up the animations, then create an animator override component, set in the Character_AnimationController, then override each animation.

This is much more convenient than duplicating the animatior controller everywhere because if we make a change then that change will be applied to each and every character using an animator controller overriding the base.